ml-model/requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# ML Libraries
joblib==1.3.2
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
lightgbm==4.1.0

# Message Queue
pika==1.3.2

# Utilities
python-dotenv==1.0.0

# Testing (optional, for test scripts)
requests==2.31.0
